Created login context for authentication
this took a lot of restructuring but it should work now
