+++ components/LoginContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export default LoginContext = createContext(null)
